Consider non-terminated instances to be existing
Currently, any non-existing instance is recreated during machine reconciliation.
In case an AWS instance is stopped for maintenance purposes, it is considered
as it would be terminated. Which is not the expected behaviour.
Instead, consider such instance to be existing until its status changes back to
running or is terminated.

Side effect: stopped worker machines will not get reconciled.