Search for existing instances by instance ID
This changes the behavior of the actuator's `getMachineInstances`
method to first look for a machine's backing instance by the instance
ID in the provider status.  If the instance ID is not set, it falls
back to the previous method of filtering by tags.  This should help
prevent issues where AWS creates the instance, but takes some time to
propagate tag values, which can result in Exists() wrongly returning
false.