pkg/actuators/machine/actuator: Drop setMachineCloudProviderSpecifics comment
The updateProviderID comment was probably a copy/paste error when
setMachineCloudProviderSpecifics landed in 716083f839 (Set additional
machine annotations/labels to get pretty machine output, 2019-07-23,
openshift/cluster-api-provider-aws#242).  I haven't read through the
function carefully enough to be able to propose a comment that adds
much beyond the function name, which is fairly clear on its own.  But
I know the updateProviderID comment is not a good fit ;).