Update Machine status with all instance IP addresses
Previously we only updated the addresses in a Machine's status with
the primary public and private IP addresses and DNS names.  That
causes problems with automated CSR renewal because the Subject
Alternative Names in CSRs come from the addresses on the Node object,
which include all addresses, e.g. ones added by attaching additional
network interfaces after an instance is created.  This changes the
behavior to match what the Kubernetes AWS cloud provider does, which
is to include the addresses of all attached interfaces.